requirements.txt
psycopg2-binary

# https://github.com/thruston/grid-banger
osgb

xlsx2csv

pytest

requests

python-statemachine[diagrams]

